move body extraction logic to HasBody trait
